requirements.txt
openai==1.58.1
typer==0.15.1
rich==13.9.4
python-dotenv==1.0.1
pynput 
pyperclip